Додано перевірку на упішність клонування репозиторію

+++ projects/make_tarball/make.sh
@@ -5,11 +5,17 @@ ARC_NAME=$SRC_DIR.tar.gz
 
 git clone git://madf.dyndns.org/stg.git $SRC_DIR
 
-rm -r $SRC_DIR/.git
+if [ $? != 0 ]
+then
+    echo "Failed to clone repository"
+    exit -1
+fi
+
+rm -rf $SRC_DIR/.git
 rm -r $SRC_DIR/projects/make_tarball
 rm -r $SRC_DIR/projects/traffcounter
 rm -r $SRC_DIR/projects/stargazer/plugins/other/userstat
 rm -r $SRC_DIR/projects/stargazer/plugins/authorization/stress
 rm -r $SRC_DIR/projects/stargazer/plugins/store/db
 
-tar -zcf $arc_name $src_dir
+tar -zcf $ARC_NAME $SRC_DIR
